Split vkubelet funcitons into separate files.
